+++ lang/spanish.php
@@ -23,6 +23,7 @@ $strings["formTemplate"] = "Plantilla de formulario";
 $strings["fillInFormInOnlyoffice"] = "Rellenar el formulario en ONLYOFFICE";
 $strings["UpdateOnlyoffice"] = "Por favor, actualice ONLYOFFICE Docs a la versión 7.0 para poder trabajar con formularios rellenables en línea";
 $strings["NotSupportedVersion"] = "Versión no compatible";
+$strings["jwt_header"] = "Encabezado de autenticación (dejar en blanco para utilizar el encabezado predeterminado)";
 $strings["connect_demo"] = "Conectarse al servidor de ONLYOFFICE Docs de demostración";
 $strings["demoPeriodIsOver"] = "El período de prueba de 30 días ha terminado, ya no puedes conectarte al servidor de ONLYOFFICE Docs de demostración";
 $strings["demoPrevMessage"] = "Este es un servidor de pruebas público, por favor no lo uses para tus datos confidenciales. El servidor estará disponible durante un período de 30 días.";