+++ lib/onlyofficePlugin.php
@@ -41,7 +41,8 @@ class OnlyofficePlugin extends Plugin implements HookPluginInterface
             [
                 "enable_onlyoffice_plugin" => "boolean",
                 "document_server_url" => "text",
-                "jwt_secret" => "text"
+                "jwt_secret" => "text",
+                "jwt_header" => "text"
             ]
         );
     }