+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This plugin is developed and maintained at https://github.com/ONLYOFFICE/onlyoff
 
 ## Added
 - Turkish and Galician empty file templates
+- jwt header setting
 
 ## Changed
 - document server v6.0 and earlier is no longer supported